@@ -6,9 +6,7 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-'use strict';
-
-/* eslint-disable no-restricted-globals */
+/* eslint-disable no-restricted-globals, @typescript-eslint/strict-boolean-expressions, @typescript-eslint/no-unnecessary-condition */
 
 const _global =
 	typeof globalThis === 'object' && isGlobal(globalThis) && globalThis ||
@@ -25,14 +23,13 @@ const _global =
 	// eslint-disable-next-line no-new-func
 	new Function('', 'return this')();
 
-module.exports = _global;
+export default _global;
 
 /**
  * Checks if the provided value is a global object by confirming the presence of Math,
  * known to exist in any global JS environment
  *
  * @param obj
- * @returns {boolean}
  */
 function isGlobal(obj) {
 	return Boolean(obj) && obj.Math === Math;
