@@ -6,15 +6,11 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-'use strict';
-
-/* eslint-disable no-var */
-
-var GLOBAL = require('core/shims/global');
+import GLOBAL from 'core/shims/global';
 
 if (typeof process === 'object' && typeof GLOBAL['requestAnimationFrame'] === 'undefined') {
 	(function requestAnimationFrameShim() {
-		GLOBAL['requestAnimationFrame'] = function requestAnimationFrame(cb) {
+		GLOBAL['requestAnimationFrame'] = function requestAnimationFrame(cb: Function) {
 			return setTimeout(() => cb(performance.now()), 0);
 		};
 
